添加TabBar中间的plus按钮

+++ MAWeibo/class/Main/controller/MWTabBarViewController.m
@@ -62,7 +62,7 @@
 /**
  *  代理方法
  */
-- (void)tabBar:(MWTabBar *)tabBar didselectButtonFrom:(int)from to:(int)to
+- (void)tabBar:(MWTabBar *)tabBar didselectButtonFrom:(NSInteger)from to:(NSInteger)to
 {
     self.selectedIndex = to;
 }
